refactor: Rename AccountUtxos -> AccountUtxoList.  Move to screens directory.
